Move build-scripts/combine-ovals.py to ssg/_build_ovals.py
